Feat: TEST_PRODUCT.ts 상수 분리 #126
